#26 - card and hub implementations completes
